requirements.txt
pandas
python-docx
PyPDF2
requests